Fix filtered noise HP/notch not working, BP distorted

+++ yarns/oscillator.h
@@ -45,7 +45,7 @@ class StateVariableFilter {
  public:
   void Init();
   void RenderInit(int16_t frequency, int16_t resonance);
-  void RenderSample(int16_t in);
+  void RenderSample(int32_t in);
   int32_t bp, lp, notch, hp;
  private:
   Interpolator<kAudioBlockSizeBits> cutoff, damp;
